+++ lib/src/checkbox/sbb_checkbox_list_item.dart
@@ -411,7 +411,7 @@ class _NonHittableCheckbox extends StatelessWidget {
             value: value,
             tristate: tristate,
             onChanged: onChanged,
-            padding: EdgeInsets.zero,
+            style: SBBCheckboxStyle(padding: EdgeInsets.zero),
             semanticLabel: semanticLabel,
           ),
         ),